Sum the partitions as long, so closure cannot be reached by overflow
Self-found while writing the round-nine review prompts. Round eight
added the invalidBreakdown closure check to stop a forged
productBodyDefect ceiling. Summed as int, that check was satisfiable by
wrapping: int.MaxValue + int.MaxValue + 9 is exactly 7, so a row
recording `invalid: 7` could state `productBodyDefect: int.MaxValue`,
close every partition, pass non-negativity, and set a ceiling no
regression could ever exceed. Verified against the real binary: that
baseline ratcheted `RATCHET OK` with exit 0.

The same wrap reached the validDifferent partition, where sub-buckets of
int.MaxValue, int.MaxValue, 51 summed to exactly 49.

Every partition sum is now computed as long, including the benchmark's
check on its own measured counts. Those are not caller-supplied, but
which side of the comparison is trusted is not a property worth
depending on.

Area=Corpus 214 -> 216, 0 failed.

Copilot-Session: 9272c125-246b-4019-943a-bfd312619af2

+++ tools/DecompilerHarness/AuthoredCorpusBenchmark.cs
@@ -344,11 +344,15 @@ static class AuthoredCorpusBenchmark
         int Unsupported,
         int UnknownOutcome)
     {
-        public int ValidDifferentSum
-            => Lowering + KnownTaste + FrontierIlExact + FrontierIlDiff + FrontierIlNoVerdict;
+        // Summed as long for the same reason the recorded-row sums are, even though
+        // these counts are measured here rather than supplied by a caller: a partition
+        // check that can wrap is not a partition check, and which side is trusted is
+        // not a property worth depending on.
+        public long ValidDifferentSum
+            => (long)Lowering + KnownTaste + FrontierIlExact + FrontierIlDiff + FrontierIlNoVerdict;
 
-        public int TopLevelSum
-            => Correct + ValidDifferent + Invalid + NotFull + Drift + Unsupported + UnknownOutcome;
+        public long TopLevelSum
+            => (long)Correct + ValidDifferent + Invalid + NotFull + Drift + Unsupported + UnknownOutcome;
 
         /// <summary>
         /// Both partitions close exactly. A shortfall means a row was counted in a
